requirements.txt
aiofiles==23.2.1
aiogram==3.6.0
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiohttp-socks==0.8.4
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.6.0
async-timeout==4.0.3
asyncio==3.4.3
asyncpg==0.29.0
attrs==23.1.0
bcrypt==4.1.3
beautifulsoup4==4.12.2
blinker==1.8.2
bs4==0.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
fake-useragent==1.5.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==4.9.3
magic-filter==1.0.12
MarkupSafe==2.1.5
MouseInfo==0.1.3
multidict==6.0.5
numexpr==2.8.7
numpy==1.26.2
outcome==1.3.0.post0
packaging==23.2
pefile==2023.2.7
Pillow==10.1.0
pluggy==1.5.0
publicip==1.0.1
pyaes==1.6.1
pyasn1==0.6.0
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==2.7.1
pydantic-settings==2.3.4
pydantic_core==2.18.2
PyGetWindow==0.0.9
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2023.10
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
pytest==8.2.0
python-dotenv==1.0.1
python-socks==2.4.4
pytweening==1.0.7
pywin32-ctypes==0.2.2
redis==5.0.4
requests==2.31.0
rsa==4.9
selenium==4.15.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.30
Telethon==1.34.0
trio==0.23.1
trio-websocket==0.11.1
typing==3.7.4.3
typing_extensions==4.11.0
urllib3==2.0.4
useragent==0.1.1
vboxapi==1.0
Werkzeug==3.0.3
wsproto==1.2.0
yarl==1.9.4
